% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{asciidoc_table}
\alias{asciidoc_table}
\title{Title}
\usage{
asciidoc_table(data.df, caption = "", the.style = c("asciimint"),
  lib = c("ascii"), be.bare = FALSE, ...)
}
\arguments{
\item{the.style}{}
}
\description{
Title
}
