% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{asciimint_table}
\alias{asciimint_table}
\title{Converts a data.frame to Asciimint versions (latex, html, asciidoc pure)}
\usage{
asciimint_table(data.df, caption = "", the.style = c("asciimint"),
  out.type = c("all", "latex", "html", "asciidoc"), html.lib = c("plotly"),
  latex.lib = c("xtable"), mixture.mode = c("sub", "sub.and.super",
  "super"), be.bare = FALSE, file.out = FALSE, file.here = FALSE,
  dirname.s, filename.s, ...)
}
\arguments{
\item{data.df}{The data.frame to create table for. If it is list of data.frames of the same size then they would be mixed regarding the mixture.mode.}

\item{...}{the lib specific parameters. If more than one lib is specified then the first one.}
}
\description{
Converts a data.frame to Asciimint versions (latex, html, asciidoc pure)
}
