% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{html_table}
\alias{html_table}
\title{Title}
\usage{
html_table(data.df, caption = "", the.style = c("asciimint"),
  lib = c("xtable"), be.bare = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
