% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{latex_table}
\alias{latex_table}
\title{Title}
\usage{
latex_table(data.df, caption = "", the.style = c("asciimint"),
  lib = c("xtable"), be.bare = FALSE, ...)
}
\arguments{
\item{caption}{}
}
\description{
Title
}
